requirements.txt
flask==3.1.1
groq==1.1.1
pypdf==5.4.0
python-docx==1.1.2
pydub==0.25.1
edge-tts==7.0.2
aiofiles==24.1.0
